moved-size-up-sprites
