+++ main.js
@@ -1,7 +1,5 @@
 import { dispatch } from "./dispatch.js";
 
-window.dispatch = dispatch;
-
 window.addEventListener("load", () => {
   dispatch("INIT");
 });
